refactored frame initializations
it was common to create a panel then put it into a frame, now we just create a frame

it was usually possible to just transform the ...Panel class into a ...Frame class. But some of the ...Panel classes were created with the intelliJ IDEA GUI design tool. So instead of figuring out if it was possible to transform those into ...Frame classes, I decided to make distinct ...Frame classes that just wrap around the generated ...Panel classes (see GroupCreationFrame and GroupManagementFrame)