refactor group deleting
now, instead of GroupManagementController taking a GroupListController to call removeGroup on when the user clicks on the "delete group" button,
it exposes a method that lets the GroupBarController do it.
But the GroupBarController doesn't really do it either -- it exposes a method that lets the GroupListController do it. Which makes sense, because deleting the gorup will involve removing the group bar from the group list, which is the responsibility of the GroupListController