refactor PopUpFrame, subclass and controllers
Instead of taking a "btnThatOpenedTheFrame" to call setEnabled(true) on it when the frame closes, now the PopUpFrame has a onClose() method that accepts a callback that is called when the frame closes.
Instead of sending the button that opened the frame as an argument to the constructor, we set the callback to () -> btnThatOpenedTheFrame.setEnabled(false).
It's not like we'll have different closing actions than re-enabling a button and need this flexibility. It just looks nicer this way.
But now the PopUpFrame is not doing much, maybe it should be deleted.