Fixed bug in the user search frame
In the user search frame, if you clicked "search" again after clicking on "ask to be friends" with some user, the "invite sent" button would turn into "ask to be friends" agian. This happened because the internal list tehe UserSearchController kept of pending friend invites didn't get updated at that "ask to be friends" button press.

An alternative fix would be to load the pendingFriendInvs at each search, but that's a more expensive operation than just updating the list.