Add InviteRepository (sql) + Invites won't have a "state" --
Also, remove InviteRepositoryInFile.
And IInviteRepository has been changed too as a consequence of removing invite states.

Invites won't have a "state" means:
If an invite exists, it's pending; when it get accepted or declined, it doesn't exist anymore, it gets deleted.
Rationale:
Suppose we store accepted and declined states in the database. Remember that they're uniquely identified by the users involved -- so if A has already asked B to be friends, he can't do it again. Now, if a friendship invite is accepted and A becomes friends with B, the invite will persist in the database with state "accepted". But what if B unfriends A? We'll need to not only remove the friendship relationship, but also the invite from A to B. Because if we don't delete the invite, A won't be able to ask B to be friends again, because the previous invite will still exist.
Of course we could just delete the invite. But at this point it's just a lot simpler and more straightforward to drop the requirement to store accepteed and declined invites.
Now, when B accepts A's invite to be friends, the invite will just be deleted, which is simpler and less messy to deal with. When B unfriends A, it's just the friendship relatonship that we have to update in the database.