Add acceptFriendInvite transactio to InviteRepository; IInviteRepository now has acceptInvite(invite)
About acceptFriendInvite:

We ignore the DELETE updated row count because:
if it's 0, ok -- somehow the invite was already deleted before we did it
if it's 1, ok -- the invite existed and we deleted it, as expected
and it can only take 0 and 1 as value, since nicknameFrom and nicknameTo
form the primary key of FriendInvites, so there shouldn't be more than
1 match.

We also ignore the INSERT updated row count because:
it's equal to 1 or 0 instead of 2
only when one or both of the (from, to) or (to, from) rows
already exists. But that already causes a SQLException and goes into the catch that calls the rollback.