Deal with updated row counts in transactions
for InviteRepository.acceptGroupInvite:
we ignore the updated row counts because for the DELETE it doesn't matter (it's 0 or 1 -- if it's 0, ok, somehow it was already deleted before we did it; if it's 1, ok, as expected) and for the INSERT the situations where the count would be different than expected are the same where the SQLException is thrown (inserting duplicate primary key)

for GroupRepository.changeOwner:
the first UPDATE only has 0 updated row count when there doesn't exist a group with the given id. But then the second update will also have updated row count of 0 because there won't exist any GroupMembership rows with groupId equal to that id, because no groups exist with that id, and groupId has a foreign key constraint. So both UPDATES will have updated row count of 0, which means nothing happened. Therefore we don't need to rollback.