Start implementing GroupChatController
To finish the implementation, deal with the edge case of a newly created group, as the comment says.

Along the way, the "messages" attribute and related methods were removed from the Group model; and the ChatFrame interface was modified.
The messages are managed in the controller itself.
Also, we don't do the unloading of messages in order to keep memory usage small because it's simpler and probably not going to be a problem if we don't do it.