Add quit button to GroupBar (and fix some responsiveness issues in it)
deleteGroup was first being called at the GroupManagementController, and then at the GroupListController, which was wrong. It was a confusion involving the onDelete method names -- GroupListController only should pass what *it* should do after a group is successfully deleted in the GroupManagementController (i.e. remove the group bar), not perform the repository deleteGroup call itself.