Some minor refactoring on the controllers
Mainly changing some methods like onDelete to afterDelete. In the GroupListController, for instance, when the GroupManagementController is created, this makes it clear that what you're passing is a callback to what should happen after the group is actually deleted, and not a callback that implements the group deletion itself -- that's what the GroupManagementController does, and if it's successful in doing so, it calls the afterDelete callback.