Minor refactoring, renaming
There was a lot of calls like showFooButton() and onClickFoo() one after the other, when they really are just one thing (you're not just gonna show the button without configuring it; you're not configure a button you're not showing). So I just merged them together (e.g. in the GroupBar, now with addQuitButton(Runnable) and addManageButton(Runnable) instead of showManageButton() then onClickManage(Runnable))